+++ src/lattice.h
@@ -1,7 +1,7 @@
 #ifndef LATTICE_H
 #define LATTICE_H
 
-int fill_lattice(int * lattice, int n, float p);
+int fill_lattice(int * lattice, int n, double p);
 int print_lattice(int * lattice, int n);
 
 #endif
